Added FireBase listing card to index.html
This commit adds a new listing card for FireBase to the index.html file. The card consists of a card-icon with the FireBase logo, and a headline-sm card-title for FireBase. The FireBase listing card is added to the existing list of cards and is placed in the correct alphabetical order.